[REFACTOR] #2 WelcomeViewController와 WelcomeView 분리
